requirements.txt
# Database
SQLAlchemy==2.0.25
mysql-connector-python==8.3.0
cloud-sql-python-connector==1.7.0
PyMySQL>=1.1.1

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
functions-framework==3.*
google-cloud-functions==1.13.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utils
python-dotenv==1.0.1
requests==2.31.0